Blotter filters (#15)
* Add PeriodFilter
* Fix tabcontrol (always close selected item)
* Change menu
* Add DuplicateCommand

+++ src/Financier.DataAccess/DataBase/Scripts/alter/20110412_0024_create_running_balance_table.sql
@@ -3,7 +3,7 @@ create table if not exists running_balance
     account_id integer not null,
     transaction_id integer not null,
     datetime long not null,
-    balance integer not null,
+    balance long not null,
     PRIMARY KEY (account_id, transaction_id)
 );
